fix some warnings

+++ gui/qt/mainwidgets/datawidget.cpp
@@ -23,7 +23,7 @@ DataWidget::~DataWidget()
 void DataWidget::updateWidget(GUI::change_t change)
 {
     if(change & GUI::Change::data){
-        if(master->data.size() > ui->DataSel->count()){
+        if(static_cast<int>(master->data.size()) > ui->DataSel->count()){
             const BaseData& dat = *master->data.back();
             ui->DataSel->addItem(dat.name.c_str());
             ui->DataSel->setCurrentIndex(ui->DataSel->count()-1);
